added hero section and responsive style
