+++ src/components/NotiomLogo.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Image } from '@chakra-ui/react';
+
+export const NotiomLogo = () => {
+  return (
+    <Image
+      boxSize='48px'
+      objectFit='cover'
+      src='Vector.png'
+    />
+  );
+};
+
